GridExecutor: added beforeDeleteOne & beforeDeleteMultiple hooks

+++ modules/GridModule/GridExecutor.class.php
@@ -731,40 +731,52 @@ MSG;
 	  //////////////////////////////////////////////////////////////////////////
 	 // DELETE
 	//////////////////////////////////////////////////////////////////////////
-
+	
+	protected function beforeDeleteMultiple($ids) {}
+	
 	public function delete_multiple() {
 		$ids = $this->request->req('ids');
 		$count = count($ids);
-		if ($count === $n = $this->table->deleteWherePkIn($ids)) {
-			$this->deletedCount = $count;
-			return true;
-		} else {
-			if ($n < $count) {
-				throw new SystemException(
-					'Delete failed',
-					lang('Une erreur a empêché la suppression de tous les enregistrements.')
-				);
-			} else if ($n > $count) {
-				Logger::getLogger('GridController')->error('Terrible mistake, I have '
-					. 'deleted more reccords than required here!!! {} rows deleted', $n);
-				throw new SystemException('Terrible Mistake');
+		if (false !== $this->beforeDeleteMultiple($ids)) {
+			if ($count === $n = $this->table->deleteWherePkIn($ids)) {
+				$this->deletedCount = $count;
+				return true;
+			} else {
+				if ($n < $count) {
+					throw new SystemException(
+						'Delete failed',
+						lang('Une erreur a empêché la suppression de tous les enregistrements.')
+					);
+				} else if ($n > $count) {
+					Logger::getLogger('GridController')->error('Terrible mistake, I have '
+						. 'deleted more reccords than required here!!! {} rows deleted', $n);
+					throw new SystemException('Terrible Mistake');
+				}
+				return false;
 			}
+		} else {
 			return false;
 		}
 	}
+	
+	protected function beforeDeleteOne($id) {}
 
 	public function delete_one() {
 		$id = $this->request->req($this->table->getPrimaryKeyName());
-		if (1 === $n = $this->table->deleteWherePkIn(array($id))) {
-			return true;
-		} else {
-			Logger::getLogger('GridController')->error('{} rows deleted', $n);
-			if ($n > 0) {
-				Logger::getLogger('GridController')->error('Terrible mistake, I have '
-					. 'deleted more than 1 reccord here!!! {} rows deleted', $n);
-				throw new SystemException('Terrible Mistake');
+		if (false !== $this->beforeDeleteOne($id)) {
+			if (1 === $n = $this->table->deleteWherePkIn(array($id))) {
+				return true;
+			} else {
+				Logger::getLogger('GridController')->error('{} rows deleted', $n);
+				if ($n > 0) {
+					Logger::getLogger('GridController')->error('Terrible mistake, I have '
+						. 'deleted more than 1 reccord here!!! {} rows deleted', $n);
+					throw new SystemException('Terrible Mistake');
+				}
+				throw new SystemException('Delete failed');
+				return false;
 			}
-			throw new SystemException('Delete failed');
+		} else {
 			return false;
 		}
 	}
